Bit of reorganizing
